requirements.txt
streamlit
tensorflow>=2.10
pillow
numpy
scikit-learn
